+++ src/index.js
@@ -17,6 +17,7 @@ import secrets from '@adobe/helix-shared-secrets';
 import dataAccess from '@adobe/spacecat-shared-data-access';
 import apex from './apex/handler.js';
 import cwv from './cwv/handler.js';
+import sitemap from './sitemap/handler.js';
 import noopHandler from './digest/handler-noop.js';
 import notFoundInternalDigestHandler from './notfound/handler-internal.js';
 import notFoundExternalDigestHandler from './notfound/handler-external.js';
@@ -27,6 +28,7 @@ import experimentation from './experimentation/handler.js';
 export const HANDLERS = {
   apex,
   cwv,
+  sitemap,
   404: noopHandler,
   '404-external': notFoundExternalDigestHandler,
   '404-internal': notFoundInternalDigestHandler,
@@ -88,9 +90,9 @@ async function run(message, context) {
 
   const handler = HANDLERS[type];
   if (!handler) {
-    const msg = `no such audit type: ${type}`;
-    log.error(msg);
-    return notFound();
+    const msg = `No handler found for the type: ${type}`;
+    log.info(msg);
+    return notFound(msg);
   }
 
   const t0 = Date.now();